% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gest.R
\name{gest}
\alias{gest}
\title{G-estimation for an end of study outcome.}
\usage{
gest(
  data,
  idvar,
  timevar,
  Yn,
  An,
  Ybin,
  Abin,
  Lny,
  Lnp,
  type = 1,
  Cn = NA,
  LnC = NA,
  ...
)
}
\arguments{
\item{data}{A data table in long format containing the data to be analysed. See description for details.}

\item{idvar}{Character string specifying the name of the ID variable in data.}

\item{timevar}{Character string specifying the name of the time variable in the data.
Note that timevar must specify time periods as integer values starting from 1 (must not begin at 0).}

\item{Yn}{Character string specifying the name of the outcome variable.}

\item{An}{Character string specifying the name of the exposure variable.}

\item{Ybin}{True or False indicator of whether the outcome is binary.}

\item{Abin}{True or False indicator of whether the exposure is binary.
Note that if \code{Abin==TRUE} then \code{An} MUST be written as a numeric variable
taking values 0 or 1. If not use \code{\link{gestcat}}}

\item{Lny}{Vector of character strings specifying the names of the
confounders to be included in the outcome model in quotations.}

\item{Lnp}{Vector of character strings specifying the names of the confounders
to be included in the model calculating the propensity scores.}

\item{type}{Value from 1-4 specifying SNMM type to fit. See details}

\item{Cn}{Optional character string specifying the name of the censoring indicator variable.
Cn should be a vector taking values 0 or 1, with 1 indicating censored.}

\item{LnC}{Vector of character strings specifying the names of the covariates to be used in the censoring score model to calculate
the censoring weights.}

\item{...}{Additional arguments, currently not in use.}
}
\value{
List of the fitted causal parameters of the posited SNMM. These are labeled as follows for each SNMM type, where An is
set to the name of the exposure variable, i is the current time period, and Lny[1] is set to the name of the first confounder in \code{Lny}.
\item{\code{type=1} }{An: The effect of exposure at any time t on outcome }
\item{\code{type=2} }{An: The effect of exposure at any time t on outcome, when Ln[1] is set to zero.\cr
An:Ln[1]: The effect modification by Lny[1], the additional effect of A on Y for each unit increase in Lny[1]. }
\item{\code{type=3} }{t=i.An: The effect of exposure at time t=i on outcome.}
\item{\code{type=4} }{t=i.An: The effect of exposure at time t=i on outcome, when Ln[1] is set to zero.\cr
t=i.An:Ln[1]: The effect modification by Lny[1], the additional effect of A on Y at time t=i for each unit increase in Lny[1].}
}
\description{
Performs G-estimation for a structural nested mean model (SNMM), based on the outcome regression methods described
in Sjolander and Vansteelendt (2016) and Dukes and Vansteelendt (2018). We assume a dataset with an end of study outcome that is either binary or continuous,
time-varying counfounding, and a time-varying exposure that is either binary or continuous.
}
\details{
Given a time-varying exposure variable, \eqn{A_t=a_t} and time-varying confounders, \eqn{L_t=l_t} measured over time periods \eqn{t=1,\ldots,T}, and an end of study outcome \eqn{Y}
measured at time \eqn{T+1}, gest estimates the causal parameters \eqn{\psi} of a SNMM of the form
 \deqn{E(Y(\bar{a}_{t},0)-Y(\bar{a}_{t-1},0)|\bar{a}_{t-1},\bar{l}_{t-1})=\psi z_ta_t \;\forall\; t=1,\ldots,T}
 if Y is continuous or
 \deqn{\frac{E(Y(\bar{a}_{t},0)|\bar{a}_{t-1},\bar{l}_{t-1})}{E(Y(\bar{a}_{t-1},0)|\bar{a}_{t-1},\bar{l}_{t-1})}=exp(\psi z_ta_t)\;\forall\; t=1,\ldots,T }
 if Y is binary. The SNMMs form is defined by the parameter \eqn{z_t}, which can be controlled by the input type as follows
 \itemize{
 \item{\code{type=1} }{Sets \eqn{z_t=1}. Now \eqn{\psi} is now the effect of exposure (or unit increase in exposure) at any time t on Y}
 \item{\code{type=2} }{Sets \eqn{z_t=c(1,l_t)}, and adds affect modification by \eqn{L_t}.
 Now \eqn{\psi=c(\psi_0,\psi_1)} where \eqn{\psi_0} is the effect of exposure (or unit increase in exposure) at any time t on Y when \eqn{l_t=0} for all t, modified by
 \eqn{\psi_1} for each unit increase in \eqn{l_t} at all times t. If there is more than one confounder, the first confounder named in Lny is taken as the effect modifier. Note that effect modification
 is currently only supported for binary (written as a numeric 0,1 vector) or continuous confounders.}
 \item {\code{type=3} }{Allows for a time-varying causal effect. Sets \eqn{z_t} to a vector of zeros of length T with a 1 in the t'th position. Now \eqn{\psi=c(\psi_1,\ldots,\psi_T)}
 where \eqn{\psi_t} is the effect of \eqn{A_t} on Y.}
 \item{\code{type=4} }{Allows for a time-varying causal effect with effect modification. Sets \eqn{z_t} to a vector of zeros of length T with \eqn{c(1,l_t)} in the t'th position.
 Now \eqn{\psi=(\underline{\psi_1},\ldots,\underline{\psi_T})} where \eqn{\underline{\psi_t}=c(\psi_{0t},\psi_{1t})}. Here \eqn{\psi_{0t}} is the effect of exposure (or unit increase in exposure) at time t on Y at \eqn{l_t=0} modified by
 \eqn{\psi_{1t}} for each unit increase in \eqn{l_t}.If there is more than one confounder, the first confounder named in Lny is taken as the effect modifier. Note that effect modification
 is currently only supported for binary (written as a numeric 0,1 vector) or continuous confounders }
 }
 The data must be in long format, where we assume the convention that each row with \code{time=t} contains \eqn{A_t,L_t} and \eqn{C_{t+1}}. That the censoring indicator for each row
 should indicate that a user is censored AFTER time t. The end of study outcome should be repeated on each row. If either A or Y are binary, they must be written as numeric vectors taking values either 0 or 1.
 The same is true for any covariate that is used for effect modification.\cr
 The data must be rectangular with a row entry for every individual for each exposure time 1 up to T. Data rows after censoring should be empty apart from the ID and time variables. This can be done using the function \code{\link{FormatData}}.\cr
 By default the censoring, propensity and outcome models include the exposure history at the previous time as a variable. One may consider also including all previous exposure and confounder history as variables in \code{Lny},\code{Lnp}, and \code{LnC} if necessary.\cr
 Censoring weights are handled as described in Sjolander and Vansteelendt (2016). Note that it is necessary that any variable included in \code{LnC} must also be in \code{Lnp}. Missing data not due to censoring is handled by removing rows with missing data prior to fitting the model. If outcome models fail to fit, consider removing covariates from \code{Lny} but keeping
 them in \code{Lnp} to reduce collinearity issues, or consider the sparseness of the data.
}
\examples{
datas<-dataexamples(n=10000,seed=123)
data=datas$datagest
idvar="id"
timevar="time"
Yn="Y"
An="A"
Ybin=FALSE
Abin=TRUE
Lny=c("L","U")
Lnp=c("L","U")
type=NA

gest(data,idvar=idvar,timevar,Yn,An,Ybin,Abin,Lny,Lnp,type=1)
gest(data,idvar=idvar,timevar,Yn,An,Ybin,Abin,Lny,Lnp,type=2)
gest(data,idvar=idvar,timevar,Yn,An,Ybin,Abin,Lny,Lnp,type=3)
gest(data,idvar=idvar,timevar,Yn,An,Ybin,Abin,Lny,Lnp,type=4)

}
\references{
Vansteelandt, S., & Sjolander, A. (2016). Revisiting g-estimation of the Effect of a Time-varying Exposure Subject to Time-varying Confounding, Epidemiologic Methods, 5(1), 37-56. doi: \url{https://doi.org/10.1515/em-2015-0005}.

Dukes, O., & Vansteelandt, S. (2018). A Note on G-Estimation of Causal Risk Ratios, American Journal of Epidemiology, 187(5), 1079–1084. doi: \url{https://doi.org/10.1093/aje/kwx347}.
}
