% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestboot.R
\name{gest.boot}
\alias{gest.boot}
\title{Percentile based bootstrap confidence intervals}
\usage{
gest.boot(
  gestfunc,
  data,
  idvar,
  timevar,
  Yn,
  An,
  Ybin,
  Abin,
  Lny,
  Lnp,
  type = 1,
  Cn = NA,
  LnC = NA,
  cutoff = NA,
  bn,
  alpha,
  onesided = "twosided",
  seed = 123,
  ...
)
}
\arguments{
\item{gestfunc}{Name (without quotations) of the gest function to run.
One of gest,gestmult,gestcat or gestmultcat.}

\item{data, idvar, timevar, Yn, An, Ybin, Abin, Lny, Lnp, type, Cn, LnC, cutoff}{Same arguments as in gest functions, to be input into gestfunc.}

\item{bn}{Number of bootstrapped datasets}

\item{alpha}{Confidence level of confidence intervals}

\item{onesided}{Controls the type of confidence interval generated. Takes one of three inputs, \code{"upper"} for upper one-sided confidence intervals,
\code{"lower"} for lower one-sided confidence intervals, and \code{"twosided"} for two-sided confidence intervals. Defaults to \code{"twosided"}.}

\item{seed}{Integer specifying the random seed for generation of bootstrap samples.}

\item{...}{additional arguments}
}
\value{
Returns a list of the following four elements.
\item{t0 }{The fitted causal parameters of the original data.}
\item{mean.boot }{The average fitted values of the causal parameters of the bootstrapped datasets.}
\item{conf }{The upper and/or lower bounds of a \eqn{1-\alpha} confidence interval for each causal parameter
comprising \eqn{\psi}.}
\item{conf.Bonferroni }{The upper and/or lower bounds of a Bonferroni corrected confidence
interval for each causal parameter comprising \eqn{psi}, used for multiple comparisons.}
}
\description{
Generates percentile based confidence intervals for the causal parameters
of a fitted SNMM by gest functions. Bonferroni corrected confidence
intervals are also reported for multiple comparisons.
}
\examples{
datas<-dataexamples(n=1000,seed=1234567)
data=datas$datagest
idvar="id"
timevar="time"
Yn="Y"
An="A"
Ybin=FALSE
Abin=TRUE
Lny=c("L")
Lnp=c("L")
gestfunc<-gest
type=NA

gest.boot(gest,data,idvar,timevar,Yn,An,Ybin,Abin,Lny,Lnp,type=1,bn=10,alpha=0.05,onesided="lower",seed=123)


}
